requirements.txt
Flask>=2.0
openai>=1.0.0
scikit-learn
pandas
joblib
reportlab
python-dotenv
validators
gunicorn
